+++ extensions/models-core/src/main/kotlin/io/medatarun/model/domain/BusinessKeyId.kt
@@ -4,4 +4,10 @@ import io.medatarun.type.commons.id.Id
 import java.util.UUID
 
 @JvmInline
-value class BusinessKeyId(override val value: UUID): Id<BusinessKeyId>
+value class BusinessKeyId(override val value: UUID): Id<BusinessKeyId> {
+    companion object {
+        fun generate(): BusinessKeyId {
+            return Id.generate(::BusinessKeyId)
+        }
+    }
+}