+++ extensions/models-core/src/main/kotlin/io/medatarun/model/domain/EntityPrimaryKeyId.kt
@@ -4,4 +4,10 @@ import io.medatarun.type.commons.id.Id
 import java.util.UUID
 
 @JvmInline
-value class EntityPrimaryKeyId(override val value: UUID): Id<EntityPrimaryKeyId>
+value class EntityPrimaryKeyId(override val value: UUID): Id<EntityPrimaryKeyId> {
+    companion object {
+        fun generate(): EntityPrimaryKeyId {
+            return Id.generate(::EntityPrimaryKeyId)
+        }
+    }
+}