+++ extensions/models-core/src/test/kotlin/io/medatarun/model/domain/ModelValidationTest.kt
@@ -13,50 +13,6 @@ import kotlin.test.assertIs
 class ModelValidationTest {
     private val validation: ModelValidation = ModelValidationImpl()
 
-    @Test
-    fun `model with bad entity identifier`() {
-        val model = ModelAggregateInMemory.builder(
-            key = ModelKey("test"),
-            version = ModelVersion("0.0.1"),
-        ) {
-            val typeStringId = TypeId.generate()
-            types = mutableListOf(
-                ModelTypeInMemory(
-                    id = typeStringId,
-                    key = TypeKey("String"),
-                    name = null,
-                    description = null
-                )
-            )
-            val goodAttributeId = AttributeId.generate()
-            val badAttributeId = AttributeId.generate()
-            val e = addEntityWithIdentifierAttributeId(
-                key = EntityKey("Contact"),
-                identifierAttributeId = badAttributeId
-            ) {
-
-            }
-            addAttribute(
-                AttributeInMemory(
-                    id = goodAttributeId,
-                    key = AttributeKey("id"),
-                    typeId = typeStringId,
-                    name = null,
-                    description = null,
-                    optional = false,
-                    tags = emptyList(),
-                    ownerId = AttributeOwnerId.OwnerEntityId(e.id)
-                )
-            )
-
-
-        }
-        val result = validation.validate(model)
-        assertIs<ModelValidationState.Error>(result)
-        kotlin.test.assertEquals(1, result.errors.size)
-        assertIs<ModelValidationErrorInvalidIdentityAttribute>(result.errors.first())
-
-    }
 
     @Test
     fun `model with bad attribute type`() {